move prevalence generating to a separate script
